Rename, move, optimize imports with new changes.
